Новая папка (10)/requirements.txt
aiogram==3.2.0
python-dotenv==1.0.0
Pillow==10.1.0
pytesseract==0.3.10
pdf2image==1.16.3
sqlite3
calamari-ocr==2.1.3
numpy==1.24.3
opencv-python==4.9.0.80
SQLAlchemy==2.0.27
python-telegram-bot==20.8 